+++ Source/Steam Library Manager/Functions/TaskManager.cs
@@ -10,15 +10,15 @@ namespace Steam_Library_Manager.Functions
 {
     internal static class TaskManager
     {
-        private static readonly NLog.Logger logger = NLog.LogManager.GetCurrentClassLogger();
+        private static readonly NLog.Logger Logger = NLog.LogManager.GetCurrentClassLogger();
 
-        public static ObservableCollection<Definitions.List.TaskInfo> TaskList = new ObservableCollection<Definitions.List.TaskInfo>();
+        public static readonly ObservableCollection<Definitions.List.TaskInfo> TaskList = new ObservableCollection<Definitions.List.TaskInfo>();
         public static CancellationTokenSource CancellationToken;
         public static bool Status, Paused, IsRestartRequired;
         public static Definitions.List.TaskInfo ActiveTask;
-        public static Definitions.List.TMInfo TMInfo { get; set; } = new Definitions.List.TMInfo();
+        public static Definitions.List.TmInfo TMInfo { get; set; } = new Definitions.List.TmInfo();
 
-        public static async Task ProcessTaskAsync(Definitions.List.TaskInfo CurrentTask)
+        private static async Task ProcessTaskAsync(Definitions.List.TaskInfo CurrentTask)
         {
             try
             {
@@ -34,10 +34,12 @@ namespace Steam_Library_Manager.Functions
                         default:
                             await CurrentTask.SteamApp.CopyFilesAsync(CurrentTask, CancellationToken.Token).ConfigureAwait(false);
                             break;
-
                         case Definitions.Enums.TaskType.Delete:
                             await CurrentTask.SteamApp.DeleteFilesAsync(CurrentTask).ConfigureAwait(false);
                             CurrentTask.SteamApp.Library.Steam.Apps.Remove(CurrentTask.SteamApp);
+                            break;
+                        case Definitions.Enums.TaskType.Compact:
+                            await CurrentTask.SteamApp.CompactTask(CurrentTask, CancellationToken.Token).ConfigureAwait(false);
                             break;
                     }
 
@@ -81,6 +83,9 @@ namespace Steam_Library_Manager.Functions
                             CurrentTask.OriginApp.DeleteFiles(CurrentTask);
 
                             CurrentTask.OriginApp.Library.Origin.Apps.Remove(CurrentTask.OriginApp);
+                            break;
+                        case Definitions.Enums.TaskType.Compact:
+                            //await CurrentTask.OriginApp.CompactTask(CurrentTask, CancellationToken.Token).ConfigureAwait(false);
                             break;
                     }
 
@@ -141,7 +146,7 @@ namespace Steam_Library_Manager.Functions
             catch (Exception ex)
             {
                 Debug.WriteLine(ex);
-                logger.Fatal(ex);
+                Logger.Fatal(ex);
             }
             finally
             {
@@ -179,7 +184,7 @@ namespace Steam_Library_Manager.Functions
                     catch (Exception ex)
                     {
                         Debug.WriteLine(ex);
-                        logger.Fatal(ex);
+                        Logger.Fatal(ex);
                     }
                 });
             }
@@ -230,7 +235,7 @@ namespace Steam_Library_Manager.Functions
             }
             catch (Exception ex)
             {
-                logger.Fatal(ex);
+                Logger.Fatal(ex);
             }
         }
 
@@ -263,7 +268,7 @@ namespace Steam_Library_Manager.Functions
             }
             catch (Exception ex)
             {
-                logger.Fatal(ex);
+                Logger.Fatal(ex);
             }
         }
 
@@ -277,7 +282,7 @@ namespace Steam_Library_Manager.Functions
             }
             catch (Exception ex)
             {
-                logger.Fatal(ex);
+                Logger.Fatal(ex);
             }
         }
 
@@ -291,7 +296,7 @@ namespace Steam_Library_Manager.Functions
             }
             catch (Exception ex)
             {
-                logger.Fatal(ex);
+                Logger.Fatal(ex);
             }
         }